+++ core/src/main/java/io/fabric8/maven/core/service/kubernetes/jib/JibBuildService.java
@@ -15,20 +15,32 @@
  */
 package io.fabric8.maven.core.service.kubernetes.jib;
 
+import com.google.cloud.tools.jib.api.ImageReference;
+import com.google.cloud.tools.jib.api.InvalidImageReferenceException;
+import com.google.cloud.tools.jib.api.JibContainerBuilder;
+import com.google.cloud.tools.jib.api.TarImage;
 import io.fabric8.maven.core.service.BuildService;
 import io.fabric8.maven.core.service.Fabric8ServiceException;
 import io.fabric8.maven.docker.config.ImageConfiguration;
+import io.fabric8.maven.docker.util.EnvUtil;
 import io.fabric8.maven.docker.util.Logger;
-import org.codehaus.plexus.component.annotations.Component;
-import org.codehaus.plexus.component.annotations.Requirement;
+import io.fabric8.maven.docker.util.MojoParameters;
+import org.apache.maven.plugin.MojoExecutionException;
 
+import java.io.IOException;
+import java.nio.file.Paths;
 import java.util.Objects;
 
-@Component(role = JibBuildService.class)
+import static io.fabric8.maven.core.service.kubernetes.jib.JibServiceUtil.addAssemblyFiles;
+import static io.fabric8.maven.core.service.kubernetes.jib.JibServiceUtil.buildContainer;
+import static io.fabric8.maven.core.service.kubernetes.jib.JibServiceUtil.containerFromImageConfiguration;
+import static io.fabric8.maven.core.service.kubernetes.jib.JibServiceUtil.imageNameFromImageConfiguration;
+
 public class JibBuildService implements BuildService {
 
-    @Requirement
-    JibAssemblyManager jibAssemblyManager;
+    private static final String TAR_SUFFIX = ".tar";
+
+    private JibAssemblyManager jibAssemblyManager;
 
     private BuildServiceConfig config;
 
@@ -44,10 +56,30 @@ public class JibBuildService implements BuildService {
     @Override
     public void build(ImageConfiguration imageConfiguration) throws Fabric8ServiceException {
        try {
-           JibServiceUtil.buildImage(config.getDockerMojoParameters(), imageConfiguration, jibAssemblyManager, log);
+           log.info("JIB image build started");
+           final JibContainerBuilder containerBuilder = containerFromImageConfiguration(imageConfiguration);
+           log.info("Preparing assembly files");
+           final TarImage tarImage = prepareAssembly(imageConfiguration, containerBuilder);
+           buildContainer(containerBuilder, tarImage, log);
+           log.info(" %s successfully built", imageNameFromImageConfiguration(imageConfiguration));
        } catch (Exception ex) {
            throw new Fabric8ServiceException("Error when building JIB image", ex);
        }
+    }
+
+    private TarImage prepareAssembly(ImageConfiguration imageConfiguration, JibContainerBuilder containerBuilder)
+      throws InvalidImageReferenceException, MojoExecutionException, IOException {
+
+        final String targetImage = imageNameFromImageConfiguration(imageConfiguration);
+        final MojoParameters mojoParameters = config.getDockerMojoParameters();
+        final String outputDir = EnvUtil.prepareAbsoluteOutputDirPath(mojoParameters, "", "").getAbsolutePath();
+        addAssemblyFiles(containerBuilder, jibAssemblyManager,
+          imageConfiguration.getBuildConfiguration().getAssemblyConfiguration(),
+          mojoParameters, targetImage, log);
+
+        final String imageTarName = ImageReference.parse(targetImage).toString().concat(TAR_SUFFIX);
+        log.info("Building Image Tarball at %s ...", imageTarName);
+        return TarImage.at(Paths.get(outputDir, imageTarName)).named(targetImage);
     }
 
     @Override